+++ wasm-runtime/polkadot/src/slicable.rs
@@ -1,5 +1,5 @@
 use runtime_support::{Vec, size_of, transmute, uninitialized, slice};
-use streamreader::Joiner;
+use joiner::Joiner;
 use endiansensitive::EndianSensitive;
 
 /// Trait that allows zero-copy read/write of value-references to/from slices in LE format.
