+++ wasm-runtime/polkadot/src/streamreader.rs
@@ -20,17 +20,6 @@ impl<'a> StreamReader<'a> {
 		Slicable::from_slice(slice)
 	}
 }
-
-pub trait Joiner {
-	fn join<T: Slicable + Sized>(self, value: &T) -> Self;
-}
-
-impl Joiner for Vec<u8> {
-	fn join<T: Slicable + Sized>(mut self, value: &T) -> Vec<u8> {
-		value.as_slice_then(|s| self.extend_from_slice(s));
-		self
-	}
-}
 /*
 // Not in use yet
 // TODO: introduce fn size_will_be(&self) -> usize; to Slicable trait and implement
